Moved repositories to correct packages. Neatened up code
