+++ src/main/java/za/ac/cput/Repository/User/ICustomerRepository.java
@@ -1,4 +1,4 @@
-package za.ac.cput.Repository;
+package za.ac.cput.Repository.User;
 
 /*
  *  Name: Tye Walker
@@ -9,9 +9,11 @@ package za.ac.cput.Repository;
  */
 
 import za.ac.cput.Entity.User.Customer;
+import za.ac.cput.Repository.IRepository;
+
 import java.util.Set;
 
-public interface ICustomerRepository extends IRepository<Customer, String>{
+public interface ICustomerRepository extends IRepository<Customer, String> {
     // Set is used because each customer is unique
     public Set<Customer> getAll();
 }
